==> 4_LDA_Analysis.ipynb <==
from ast import literal_eval
from gensim.corpora.dictionary import Dictionary
from gensim.models import LdaModel, ldamodel, LdaMulticore
from gensim.parsing.preprocessing import STOPWORDS
from gensim.test.utils import common_corpus, common_texts, common_dictionary
from gensim.utils import simple_preprocess
from nltk import WordNetLemmatizer
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer
from nltk.stem.porter import *
from nltk.stem.snowball import SnowballStemmer
from nltk.tokenize import sent_tokenize, word_tokenize
from sklearn.decomposition import LatentDirichletAllocation
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import GridSearchCV
from statistics import mean
import gensim
import nltk
import numpy as np
import os
import pandas as pd
import random
import warnings
warnings.filterwarnings("ignore")


stemmer = SnowballStemmer("english")
lemmatizer = WordNetLemmatizer()

overview = pd.read_csv("debate_data/data_overviews/overview_1.csv")

def string_2_row(df_w_col_name):
    count = 0
    for row in df_w_col_name:
        df_w_col_name[count] = literal_eval(row)
        count+=1
        
string_2_row(overview["moderator"])
string_2_row(overview["participants"])

# create stop_words list to be removed when LDA analysis is being conducted

nltk_stop = list(set(stopwords.words('english')))

political_stop_words = ["president", "american", "america", "govern", "government", "campaign",\
                        "country","applause","question","governor", "congresswoman", "congress",\
                       "congressman", "partisan", "republican", "democrat", "DNC","GOP","liberal",\
                       "conservative", "senate", "state", "left","right", "nomination","parti","party",
                       "democraci","constitut","politician","political","politics","pols","politi","federal",\
                       "feder", "polici","state","minister","govern","caucus","primary"]
                       
other_stop_words = ["think", "thank", "you","would", "peopl","work","reform", "family", "administr",\
                    "program","thing","peopl", "support","person","progress",\
                    "nominate""nomination","court","argument","process","man","men","woman","women", "legisl",\
                   "languag", "respect","report", "place", "today","title","issu","issue","titl", "year",\
                   "system", "record","world","laughter","laugh","farmer","earmark","office","offic",
                   "power","church", "someth", "reason", "promise", "promis", "tonight","everybodi","formosa",
                   "someon", "creat", "simpl", "percent", "faith", "mean", "need", "pocket" "percent", \
                    "secretari", "point", "noth", "ground", "build", "public", "middl", "mean", "centuri",\
                    "minut", "becom", "answer", "matter", "folk", "cours", "mean", "order" , "percent",\
                    "term", "posit", "communiti", "health", "mandat", "children", "public", "speech", \
                    "nomine", "differ", "month", "point", "everyon", "stand","everi","believ", "opportun",\
                   "understand","bring","choice", "histori", "servic", "background", "check", "nation", \
                    "littl", "friend", "critic","inequ", "realli", "somebodi", "problem", "share", "challeng", \
                    "togeth", "street", "start", "continu", "improv", "trust", "provid", "decis"]

names = ["michael", "donald", "chris", 'hillary', 'david', 'martha', 'hillari' ,"chang","barack","putin","jeb"]

for mods in overview["moderator"]:
    try:
        for mod in mods:
            names.append(mod)
    except:
        pass

for parts in overview["participants"]:
    try:
        for part in parts:
            names.append(part[:-4])
    except:
        pass

for title in overview["debate_name"]:
    t = title.replace("-"," ").split(" ")
    for z in t:
        names.append(str(z))

names = list(set(names))


states = ['Alabama', 'Alaska', 'Arizona', 'Arkansas', 'California', 'Colorado', 'Connecticut', \
          'Delaware', 'District of Columbia', 'Florida', 'Georgia', 'Hawaii', 'Idaho', 'Illinois', \
          'Indiana', 'Iowa', 'Kansas', 'Kentucky', 'Louisiana', 'Maine', 'Montana', 'Nebraska', 'Nevada', \
          'New Hampshire', 'New Jersey', 'New Mexico', 'New York', 'North Carolina', 'North Dakota', 'Ohio', \
          'Oklahoma', 'Oregon', 'Maryland', 'Massachusetts', 'Michigan', 'Minnesota', 'Mississippi', 'Missouri',\
          'Pennsylvania', 'Rhode Island', 'South Carolina', 'South Dakota', 'Tennessee', 'Texas', 'Utah', 
          'Vermont', 'Virginia', 'Washington', 'West Virginia', 'Wisconsin', 'Wyoming']

splitted_states = []
for s in states:
    for x in s.split(" "):
        splitted_states.append(x)
splitted_states = list(set(splitted_states))

stop_words = list(set(nltk_stop + political_stop_words + other_stop_words + names + splitted_states))

stop_words = [lemmatizer.lemmatize(t) for t in stop_words]
stop_words = [stemmer.stem(t) for t in stop_words]
stop_words = [t.lower() for t in stop_words]

stop_words = list(set(stop_words))

# function below finds the best params for the LDA model

def best_params(filename):
    
    try:
        df = pd.read_csv(directory+"/"+filename)
    except:
        df = filename
        
    processed_docs = []
    
    for row in df["speech"]:
        sentence = []
        for word in row.split(" "):
            x = re.sub('[^a-zA-Z]+', '', word)
            if len(x) > 4:
                if str(nltk.pos_tag([x])[0][1])[0] in ["N","J"]:
                    if x not in stop_words:
                        sentence.append(x)
        processed_docs.append(sentence)
        
    flat_list = []
    for sublist in processed_docs:
        for item in sublist:
            flat_list.append(item)

    vectorizer = TfidfVectorizer(analyzer='word', min_df=10, stop_words='english', lowercase=True,\
                                 token_pattern='[a-zA-Z0-9]{4,}')
    data_vectorized = vectorizer.fit_transform(flat_list)
    
    # Define Search Param
    search_params = {'n_components': [8,10,12,14,15],
                     'learning_decay': [0.6, 0.7, 0.8, 0.9]}

    # Initiate the Model
    lda = LatentDirichletAllocation()

    # Init Grid Search Class
    model = GridSearchCV(lda, param_grid = search_params)

    # Do the Grid Search
    model.fit(data_vectorized)
    
    try:
        best_lda_model = model.best_estimator_
        print("Best Model's Params: ", model.best_params_)
        print("Best Log Likelihood Score: ", model.best_score_)
    except:
        pass

    return (filename, model.best_params_)

# total topics

all_speech = []
directory = "debate_data/debates_part"
for filename in os.listdir(directory):
    if filename.endswith(".csv"):
        df = pd.read_csv(directory+"/"+filename)
        for row in df["speech"]:
            all_speech.append(row)

            
df = pd.DataFrame(all_speech, columns = ["speech"])         
            
x = best_params(df)

# best params for debates

samples = []

while(list(set(samples))) < ((0.25)*(len(overview["coder"]))):
    samples.append(overview["coder"][random.randint(0, len(overview["coder"]))])    

print(samples)
    
directory = "debate_data/debates_part"

best_param_results = []
counter = 1
for filename in os.listdir(directory):
    
    if filename.endswith(".csv"):
        
        coder = str(filename.split(".")[0])
        
        if coder in samples:

            print("number ", counter, ":", coder)
            counter+=1
            x = best_params(filename)
            best_param_results.append(x)
            
print("RESULTS FOR DEBATE TOPICS\n")
ds = []
for i in range(len(best_param_results)):
    ds.append(best_param_results[i][1])

d = {}
for i in ds:
    for k in i.keys():
        d[k] = tuple(d[k] for d in ds)

for st,vals in d.items():
    print("Average value for {} is {}".format(st,mean(vals)))
    try:
        print("Std for {} is {}".format(st,np.std(vals)))
    except:
        pass

directory = "debate_data/debates_part"

for filename in os.listdir(directory):
    if filename.endswith(".csv"):
        df = pd.read_csv(directory+"/"+filename)
        coder = str(filename.split(".")[0])

        try:

            processed_docs = []

            for row in df["speech"]:

                sentence = []
                for word in row.split(" "):
                    x = re.sub('[^a-zA-Z]+', '', word)
                    if len(x) > 4:
                        if str(nltk.pos_tag([x])[0][1])[0] in ["N","J"]:
                            if x not in stop_words:
                                sentence.append(x)
                processed_docs.append(sentence)

            dictionary = gensim.corpora.Dictionary(processed_docs)
            dictionary.filter_extremes(no_above = 0.75)
            bow_corpus = [dictionary.doc2bow(doc) for doc in processed_docs]

            lda_model = LdaMulticore(bow_corpus,
                      num_topics = 4,
                      id2word = dictionary,
                      workers = 3,
                      passes = 15,
                      decay = 0.77,
                      minimum_probability = 0.0)        

            top_topics = lda_model.get_document_topics(bow_corpus, minimum_probability=0.0)

            topic_n = []
            topic_share = []
            for i in range(4):
                topic_n.append(i)
                count = 0
                n = 0
                for x in top_topics:
                    count += x[i][1]
                    n += 1
                avg = count / n
                topic_share.append(avg)

            sum_terms = []
            topic_terms = []

            for i in lda_model.print_topics(num_words = 8):
                words = []
                value = 0
                out_list = i[1].split("+")
                for x in out_list:
                    num = literal_eval(x.split("*")[0][1:])
                    value+=num

                sum_terms.append(value)

                for x in out_list:
                    t = x.replace('"', " ").strip(" ").split("*")[1]
                    words.append(t.strip(" "))
                stringed = ', '.join(words)
                topic_terms.append([stringed])

            df = pd.DataFrame(list(zip(topic_share, topic_terms, sum_terms)),
                              columns = ["topic_share", "topic_terms","sum_terms"]) 

            print(coder, "\n", df ,"\n")
            
            ###SAVE LDA MODEL UNIQUE FOR EACH DEBATE
            
        except:
            print("\n********", coder, " failed to be modelled \n")

from gensim.test.utils import common_corpus, common_dictionary

lda = LdaMulticore(common_corpus,
                   id2word = common_dictionary,
                   num_topics = 5,
                   workers = 3,
                   passes = 10,
                   decay = 0.77)


def processed_docs(filename, count):
        
    df = pd.read_csv(directory+"/"+filename)
    
    processed_docs = []
    
    for row in df["speech"]:
        sentence = []
        for word in row.split(" "):
            x = re.sub('[^a-zA-Z]+', '', word)
            if len(x) > 4:
                if str(nltk.pos_tag([x])[0][1])[0] in ["N","J"]:
                    if x not in stop_words:
                        sentence.append(x)
        processed_docs.append(sentence)
        
    
    dictionary = gensim.corpora.Dictionary(processed_docs)
    dictionary.filter_extremes(no_above = 0.8)
    bow_corpus = [dictionary.doc2bow(doc) for doc in processed_docs]

    for i in range(len(rev_train)):
        top_topics = lda_train.get_document_topics(train_corpus[i], minimum_probability=0.0)
        topic_vec = [top_topics[i][1] for i in range(20)]
        topic_vec.extend([rev_train.iloc[i].real_counts]) # counts of reviews for restaurant
        topic_vec.extend([len(rev_train.iloc[i].text)]) # length review
        train_vecs.append(topic_vec)

    
    
    #lda_model = LdaMulticore(bow_corpus,
    #                 num_topics = 5,
    #                 id2word = dictionary,
    #                 workers = 3,
    #                 passes = 10,
    #                 decay = 0.74,
    #                 minimum_probability = 0.01)        
        
    
    topic_value = []
    for i in lda_model.print_topics():
        words = []
        value = 0
        out_list = i[1].split("+")
        for x in out_list:
            num = literal_eval(x.split("*")[0][1:])
            value+=num
        for x in out_list:
            t = x.replace('"', " ").strip(" ").split("*")[1]
            words.append(t.strip(" "))
        stringed = ','.join(words)
        value = round(value, 5)
        topic_value.append((value, stringed))
    
    topic_value = topic_value.sort(reverse=True)
    
    return topic_value

directory = "debate_data/debates_preprocessed"

iteration = 0
details = []
for filename in os.listdir(directory):
    this_file = []
    if filename.endswith(".csv"):
        print(filename)
        this_file.append(filename)
        result = processed_docs(filename, iteration)
        this_file.append(result)
        for i in result:
            print(i)
    details.append(this_file)
    iteration+=1

# manually inspect each topic and translate them to a new file
